+++ cmd/sqlpilot/main.go
@@ -10,7 +10,7 @@ import (
 )
 
 var (
-	version = "0.3.0"
+	version = "0.4.0"
 )
 
 func main() {